More changes and compilation stuff

+++ fclone.py
@@ -1,11 +1,12 @@
 import sys
 import os
+import os.path
 from tokeniser import tokenise
 from detector import compare
 
 def start():
 	if sys.version_info[0] >= 3:
-		print ("Error: This program requires Python 2.7.x")
+		print ("Error: This program is incompatible with Python 3.x")
 		sys.exit()
 	if len(sys.argv) != 4:
 		print "Error: Incorrect number of arguments. Please refer to README.md for more details"
@@ -24,10 +25,14 @@ def start():
 	compare(file1+"t", file2+"t", int(substring))
 
 	# clean intermediate files
-	os.remove(file1+"p")
-	os.remove(file1+"t")
-	os.remove(file2+"p")
-	os.remove(file2+"t")
+	if os.path.isfile(file1+"p"):
+		os.remove(file1+"p")
+	if os.path.isfile(file1+"t"):
+		os.remove(file1+"t")
+	if os.path.isfile(file2+"p"):
+		os.remove(file2+"p")
+	if os.path.isfile(file2+"t"):
+		os.remove(file2+"t")
 
 	#print "done"
